Fix empty line issue and some minor issues
Line was defined to be not empty, because it needed to consider EOF: if it were defined otherwise, then expressions like Line* would not be allowed, because they could represent empty strings. The simplest solution was to take care, that there is always an empty line at the end, so that Line can always end with \n, a non-empty character.